File name cleanup. Removed the redundant metrics (for example: 'metricsAppServerList' becomes appServerList and metricsAlerts becomes alerts. Moved ServerTypes into metricsTypes file. Renamed metricsAvailability to urlAvailabilityDetails and urlResponseTimeDetails for more intuitive names.
