+++ console/src/main/scripts/plugins/metrics/ts/appServerDetails.ts
