+++ console/src/main/scripts/plugins/metrics/ts/appServerList.ts
