+++ console/src/main/scripts/plugins/metrics/ts/metricsTypes.ts
@@ -218,4 +218,32 @@ module HawkularMetrics {
     alertType:string;
   }
 
+  export class ServerStatus {
+
+    constructor(public value:string, public state:string, public icon:string) {
+    }
+
+    static SERVER_UP = new ServerStatus('Up', 'up', 'fa-arrow-up');
+    static SERVER_DOWN = new ServerStatus('Down', 'down', 'fa-arrow-down');
+    static SERVER_UNKNOW = new ServerStatus('Unknown', 'unknown', 'fa-chain-broken');
+    static SERVER_STARTING = new ServerStatus('Starting', 'starting', 'fa-spinner');
+    static SERVER_RESTART_REQUIRED = new ServerStatus('Restart Required', 'restart required', 'fa-repeat');
+
+    toString = () => {
+      return this.value;
+    };
+  }
+
+  export class ServerType {
+    constructor(public value:string, public type:string) {
+    }
+
+    static SERVER_EAP = new ServerType('EAP', 'eap');
+    static SERVER_WILDFLY = new ServerType('WildFly', 'wildfly');
+
+    toString = () => {
+      return this.value;
+    };
+  }
+
 }
