+++ console/src/main/scripts/plugins/metrics/ts/urlAvailabilityDetails.ts
