+++ console/src/main/scripts/plugins/metrics/ts/urlResponseTimeDetails.ts
